requirements.txt
django==2.1.4
graphene-django==2.2.0
django-filter==2.0.0
django-graphql-jwt==0.1.5
ChatterBot==1.0.5
chatterbot-corpus==1.2.0
scikit-learn==0.20.3